fix(contoller): controller 구조 변경 (#5)
