+++ src/main/java/com/dondoc/controller/FarmController.java
@@ -1,5 +1,6 @@
 package com.dondoc.controller;
 
+import com.dondoc.dto.FarmMembers;
 import com.dondoc.dto.Farms;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -10,8 +11,8 @@ import java.util.Arrays;
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/farms")
-public class FarmsController {
+@RequestMapping("/api/farm")
+public class FarmController {
 
     @GetMapping
     public List<Farms> getFarms() {
@@ -19,6 +20,17 @@ public class FarmsController {
                 new Farms(1L, "KB농장", LocalDateTime.now()),
                 new Farms(2L, "신한농장", LocalDateTime.now()),
                 new Farms(3L, "우리농장", LocalDateTime.now())
+        );
+    }
+
+    @GetMapping("/members")
+    public List<FarmMembers> getFarmMembers() {
+        return Arrays.asList(
+                new FarmMembers(1L, 1L, 1L, LocalDateTime.now()),
+                new FarmMembers(2L, 2L, 1L, LocalDateTime.now()),
+                new FarmMembers(3L, 3L, 2L, LocalDateTime.now()),
+                new FarmMembers(4L, 1L, 2L, LocalDateTime.now()),
+                new FarmMembers(5L, 4L, 3L, LocalDateTime.now())
         );
     }
 }
